Basic functionality.
Turns on and off lights based on motion.

+++ ci.sh
@@ -3,12 +3,16 @@
 set -euo pipefail
 
 pushd motion-sensor
+constants_exists=
 
 if [ ! -f "lib/constants/constants.h" ]; then
   cp lib/constants/constants.sample.h lib/constants/constants.h
+  constants_exists=true
 fi
 
 platformio run
-rm lib/constants/constants.h
+if [ "$constants_exists" = true]; then
+  rm lib/constants/constants.h
+fi
 
 popd
